Assume flow entry is always present in hashmap
The entry(..).or_insert call(obj::new()) has the drawback of calling the
object constructor every time, even if already present in hashmap.
If the plugin constructor takes time, this slows down a lot the runtime.

Assume the object is already present (using expect) to speed up
functions. This will panic if flow is not present in hashmap.